refactor: 2단계 지원서 파일 훅 분리
